rest - avro profile refactoring.
